--- nipoppy/pipeline_store/__init__.py
@@ -1,1 +1,0 @@
-from .validation import check_pipeline_bundle
