+++ nipoppy/workflows/pipeline_store/validate.py
@@ -4,7 +4,7 @@ import logging
 from pathlib import Path
 
 from nipoppy.env import LogColor, StrOrPathLike
-from nipoppy.pipeline_store.validation import check_pipeline_bundle
+from nipoppy.pipeline_validation import check_pipeline_bundle
 from nipoppy.workflows.base import BaseWorkflow
